issue #99] - moved 'select and highlight element' function to a mixin
